combine Vue-storefront examples, use new registerComponent
