+++ examples/vue/vue-storefront-2/pages/_path.vue
@@ -1,27 +1,11 @@
 <template>
   <div id="home">
-    <div>Hello world from your Vue project. Below is Builder Content:</div>
-
-    <div v-if="canShowContent">
-      <div>page: {{ (content && content.data && content.data.title) || 'Unpublished' }}</div>
-      <builder-render-content
-        model="page"
-        :content="content"
-        :api-key="apiKey"
-        :customComponents="getRegisteredComponents()"
-      />
-    </div>
+    <BuilderContent />
   </div>
 </template>
 <script>
 import Vue from 'vue';
 import cacheControl from '../helpers/cacheControl';
-
-import { REGISTERED_COMPONENTS } from './init-builder';
-import { getContent, isEditing, isPreviewing, getBuilderSearchParams } from '@builder.io/sdk-vue';
-
-// TODO: enter your public API key
-const BUILDER_PUBLIC_API_KEY = 'f1a790f8c3204b3b8c5c1795aeac4660'; // ggignore
 
 export default Vue.extend({
   name: 'Home',
@@ -29,38 +13,6 @@ export default Vue.extend({
     'max-age': 60,
     'stale-when-revalidate': 5,
   }),
-  data: () => ({
-    canShowContent: false,
-    content: null,
-    apiKey: BUILDER_PUBLIC_API_KEY,
-  }),
-  methods: {
-    getRegisteredComponents() {
-      return REGISTERED_COMPONENTS;
-    },
-  },
-  mounted() {
-    // we need to re-run this check on the client in case of SSR
-    this.canShowContent = this.content || isEditing() || isPreviewing();
-  },
-  async fetch() {
-    const content = await getContent({
-      model: 'page',
-      apiKey: BUILDER_PUBLIC_API_KEY,
-      options: getBuilderSearchParams(this.$route.query),
-      userAttributes: {
-        urlPath: this.$route.path,
-      },
-    });
-    this.canShowContent = content || isEditing();
-    this.content = content;
-
-    if (!this.canShowContent) {
-      if (this.$nuxt.context?.ssrContext?.res) {
-        this.$nuxt.context.ssrContext.res.statusCode = 404;
-      }
-    }
-  },
 });
 </script>
